@@ -11,7 +11,7 @@ algorithm.
    https://gitlab.com/ianjcalvert/edgehammer
 """
 
-from __future__ import absolute_import, division, print_function
+from __future__ import annotations
 
 from math import ceil
 
