@@ -1,4 +1,4 @@
-from __future__ import absolute_import, division
+from __future__ import annotations
 
 import pandas as pd
 
@@ -13,8 +13,8 @@ __all__ = ()
 
 
 @bypixel.pipeline.register(pd.DataFrame)
-def pandas_pipeline(df, schema, canvas, glyph, summary):
-    return glyph_dispatch(glyph, df, schema, canvas, summary)
+def pandas_pipeline(df, schema, canvas, glyph, summary, *, antialias=False):
+    return glyph_dispatch(glyph, df, schema, canvas, summary, antialias=antialias)
 
 
 glyph_dispatch = Dispatcher()
@@ -23,14 +23,16 @@ glyph_dispatch = Dispatcher()
 @glyph_dispatch.register(_PointLike)
 @glyph_dispatch.register(_GeometryLike)
 @glyph_dispatch.register(_AreaToLineLike)
-def default(glyph, source, schema, canvas, summary, cuda=False):
-    create, info, append, _, finalize = compile_components(summary, schema, glyph, cuda)
+def default(glyph, source, schema, canvas, summary, *, antialias=False, cuda=False):
+    create, info, append, _, finalize, antialias_stage_2 = compile_components(
+        summary, schema, glyph, antialias=antialias, cuda=cuda)
     x_mapper = canvas.x_axis.mapper
     y_mapper = canvas.y_axis.mapper
-    extend = glyph._build_extend(x_mapper, y_mapper, info, append)
+    extend = glyph._build_extend(x_mapper, y_mapper, info, append, antialias_stage_2)
 
     x_range = canvas.x_range or glyph.compute_x_bounds(source)
     y_range = canvas.y_range or glyph.compute_y_bounds(source)
+    canvas.validate_ranges(x_range, y_range)
 
     width = canvas.plot_width
     height = canvas.plot_height
@@ -42,6 +44,7 @@ def default(glyph, source, schema, canvas, summary, cuda=False):
     y_axis = canvas.y_axis.compute_index(y_st, height)
 
     bases = create((height, width))
+
     extend(bases, source, x_st + y_st, x_range + y_range)
 
     return finalize(bases,
