@@ -1,4 +1,4 @@
-from __future__ import absolute_import
+from __future__ import annotations
 from datashader.glyphs.quadmesh import _QuadMeshLike
 from datashader.data_libraries.pandas import default
 from datashader.core import bypixel
@@ -15,13 +15,15 @@ glyph_dispatch = Dispatcher()
 
 
 @bypixel.pipeline.register(xr.Dataset)
-def xarray_pipeline(xr_ds, schema, canvas, glyph, summary):
+def xarray_pipeline(xr_ds, schema, canvas, glyph, summary, *, antialias=False):
     cuda = cupy and isinstance(xr_ds[glyph.name].data, cupy.ndarray)
     if not xr_ds.chunks:
-        return glyph_dispatch(glyph, xr_ds, schema, canvas, summary, cuda)
+        return glyph_dispatch(
+            glyph, xr_ds, schema, canvas, summary, antialias=antialias, cuda=cuda)
     else:
         from datashader.data_libraries.dask_xarray import dask_xarray_pipeline
-        return dask_xarray_pipeline(glyph, xr_ds, schema, canvas, summary, cuda)
+        return dask_xarray_pipeline(
+            glyph, xr_ds, schema, canvas, summary, antialias=antialias, cuda=cuda)
 
 
 # Default to default pandas implementation
