@@ -1,9 +1,9 @@
-from __future__ import absolute_import
+from __future__ import annotations
 
 import re
 
-from distutils.version import LooseVersion
 from functools import total_ordering
+from packaging.version import Version
 
 import numpy as np
 import pandas as pd
@@ -336,7 +336,7 @@ Cannot check equality of RaggedArray values of unequal length
                 self.start_indices, self.flat_array,
                 other.start_indices, other.flat_array)
         else:
-            # Convert other to numpy arrauy
+            # Convert other to numpy array
             if not isinstance(other, np.ndarray):
                 other_array = np.asarray(other)
             else:
@@ -449,7 +449,7 @@ Cannot check equality of RaggedArray of length {ra_len} with:
                 # check for NA values
                 isna = pd.isna(item)
                 if isna.any():
-                    if LooseVersion(pd.__version__) > '1.0.1':
+                    if Version(pd.__version__) > Version('1.0.1'):
                         item[isna] = False
                     else:
                         raise ValueError(
