@@ -1,12 +1,13 @@
-from __future__ import absolute_import, division, print_function
-
+from __future__ import annotations
+from packaging.version import Version
 import numpy as np
 from datashape import dshape, isnumeric, Record, Option
 from datashape import coretypes as ct
 from toolz import concat, unique
 import xarray as xr
 
-from datashader.glyphs.glyph import isnull
+from datashader.enums import AntialiasCombination
+from datashader.utils import isnull
 from numba import cuda as nb_cuda
 
 try:
@@ -17,10 +18,11 @@ except ImportError:
 
 try:
     import cudf
+    import cupy as cp
 except Exception:
-    cudf = None
+    cudf = cp = None
 
-from .utils import Expr, ngjit, nansum_missing
+from .utils import Expr, ngjit, nansum_missing, nanmax_in_place, nansum_in_place
 
 
 class Preprocess(Expr):
@@ -37,12 +39,13 @@ class extract(Preprocess):
     """Extract a column from a dataframe as a numpy array of values."""
     def apply(self, df):
         if cudf and isinstance(df, cudf.DataFrame):
-            import cupy
             if df[self.column].dtype.kind == 'f':
                 nullval = np.nan
             else:
                 nullval = 0
-            return cupy.array(df[self.column].to_gpu_array(fillna=nullval))
+            if Version(cudf.__version__) >= Version("22.02"):
+                return df[self.column].to_cupy(na_value=nullval)
+            return cp.array(df[self.column].to_gpu_array(fillna=nullval))
         elif isinstance(df, xr.Dataset):
             # DataArray could be backed by numpy or cupy array
             return df[self.column].data
@@ -68,8 +71,18 @@ class CategoryPreprocess(Preprocess):
         """Applies preprocessor to DataFrame and returns array"""
         raise NotImplementedError("apply not implemented")
 
+
 class category_codes(CategoryPreprocess):
-    """Extract just the category codes from a categorical column."""
+    """
+    Extract just the category codes from a categorical column.
+
+    To create a new type of categorizer, derive a subclass from this
+    class or one of its subclasses, implementing ``__init__``,
+    ``_hashable_inputs``, ``categories``, ``validate``, and ``apply``.
+
+    See the implementation of ``category_modulo`` in ``reductions.py``
+    for an example.
+    """
     def categories(self, input_dshape):
         return input_dshape.measure[self.column].categories
 
@@ -81,6 +94,8 @@ class category_codes(CategoryPreprocess):
 
     def apply(self, df):
         if cudf and isinstance(df, cudf.DataFrame):
+            if Version(cudf.__version__) >= Version("22.02"):
+                return df[self.column].cat.codes.to_cupy()
             return df[self.column].cat.codes.to_gpu_array()
         else:
             return df[self.column].cat.codes.values
@@ -113,7 +128,9 @@ class category_modulo(category_codes):
 
     def apply(self, df):
         result = (df[self.column] - self.offset) % self.modulo
-        if cudf and isinstance(df, cudf.DataFrame):
+        if cudf and isinstance(df, cudf.Series):
+            if Version(cudf.__version__) >= Version("22.02"):
+                return result.to_cupy()
             return result.to_gpu_array()
         else:
             return result.values
@@ -152,15 +169,20 @@ class category_binning(category_modulo):
 
     def apply(self, df):
         if cudf and isinstance(df, cudf.DataFrame):
-            ## dunno how to do this in CUDA
-            raise NotImplementedError("this feature is not implemented in cuda")
+            if Version(cudf.__version__) >= Version("22.02"):
+                values = df[self.column].to_cupy(na_value=cp.nan)
+            else:
+                values = cp.array(df[self.column].to_gpu_array(fillna=True))
+            nan_values = cp.isnan(values)
         else:
-            value = df[self.column].values
-            index = ((value - self.bin0) / self.binsize).astype(int)
-            index[index < 0] = self.bin_under
-            index[index >= self.nbins] = self.bin_over
-            index[np.isnan(value)] = self.nbins
-            return index
+            values = df[self.column].to_numpy()
+            nan_values = np.isnan(values)
+
+        index = ((values - self.bin0) / self.binsize).astype(int)
+        index[index < 0] = self.bin_under
+        index[index >= self.nbins] = self.bin_over
+        index[nan_values] = self.nbins
+        return index
 
 
 class category_values(CategoryPreprocess):
@@ -193,12 +215,14 @@ class category_values(CategoryPreprocess):
             else:
                 nullval = 0
             a = cupy.asarray(a)
-            b = cupy.asarray(df[self.column].to_gpu_array(fillna=nullval))
+            if Version(cudf.__version__) >= Version("22.02"):
+                b = df[self.column].to_cupy(na_value=nullval)
+            else:
+                b = cupy.asarray(df[self.column].fillna(nullval))
             return cupy.stack((a, b), axis=-1)
         else:
             b = df[self.column].values
             return np.stack((a, b), axis=-1)
-
 
 
 class Reduction(Expr):
@@ -212,12 +236,21 @@ class Reduction(Expr):
         if not isnumeric(in_dshape.measure[self.column]):
             raise ValueError("input must be numeric")
 
-    def out_dshape(self, in_dshape):
-        return self._dshape
-
     @property
     def inputs(self):
         return (extract(self.column),)
+
+    def _antialias_requires_2_stages(self):
+        # Return True if this Reduction must be processed with 2 stages,
+        # False if it doesn't matter.
+        # Overridden in derived classes as appropriate.
+        return False
+
+    def _antialias_stage_2(self, self_intersect, array_module):
+        # Only called if using antialiased lines. Overridden in derived classes.
+        # Returns a tuple containing an item for each constituent reduction.
+        # Each item is (AntialiasCombination, zero_value)).
+        raise NotImplementedError(f"{type(self)}._antialias_stage_2 is not defined")
 
     def _build_bases(self, cuda=False):
         return (self,)
@@ -225,26 +258,76 @@ class Reduction(Expr):
     def _build_temps(self, cuda=False):
         return ()
 
-    def _build_create(self, dshape):
-        return self._create
+    def _build_create(self, required_dshape):
+        fields = getattr(required_dshape.measure, "fields", None)
+        if fields is not None and len(required_dshape.measure.fields) > 1:
+            # If more than one field then they all have the same dtype so can just take the first.
+            first_field = required_dshape.measure.fields[0]
+            required_dshape = dshape(first_field[1])
 
-    def _build_append(self, dshape, schema, cuda=False):
+        if isinstance(required_dshape, Option):
+            required_dshape = dshape(required_dshape.ty)
+
+        if required_dshape == dshape(ct.bool_):
+            return self._create_bool
+        elif required_dshape == dshape(ct.float32):
+            return self._create_float32_nan
+        elif required_dshape == dshape(ct.float64):
+            return self._create_float64_nan
+        elif required_dshape == dshape(ct.uint32):
+            return self._create_uint32
+        else:
+            raise NotImplementedError(f"Unexpected dshape {dshape}")
+
+    def _build_append(self, dshape, schema, cuda, antialias, self_intersect):
         if cuda:
-            if self.column is None:
+            if antialias and self.column is None:
+                return self._append_no_field_antialias_cuda
+            elif antialias:
+                return self._append_antialias_cuda
+            elif self.column is None:
                 return self._append_no_field_cuda
             else:
                 return self._append_cuda
         else:
-            if self.column is None:
+            if antialias and self.column is None:
+                return self._append_no_field_antialias
+            elif antialias:
+                return self._append_antialias
+            elif self.column is None:
                 return self._append_no_field
             else:
                 return self._append
 
-    def _build_combine(self, dshape):
+    def _build_combine(self, dshape, antialias):
         return self._combine
 
     def _build_finalize(self, dshape):
         return self._finalize
+
+    @staticmethod
+    def _create_bool(shape, array_module):
+        return array_module.zeros(shape, dtype='bool')
+
+    @staticmethod
+    def _create_float32_nan(shape, array_module):
+        return array_module.full(shape, array_module.nan, dtype='f4')
+
+    @staticmethod
+    def _create_float64_nan(shape, array_module):
+        return array_module.full(shape, array_module.nan, dtype='f8')
+
+    @staticmethod
+    def _create_float64_empty(shape, array_module):
+        return array_module.empty(shape, dtype='f8')
+
+    @staticmethod
+    def _create_float64_zero(shape, array_module):
+        return array_module.zeros(shape, dtype='f8')
+
+    @staticmethod
+    def _create_uint32(shape, array_module):
+        return array_module.zeros(shape, dtype='u4')
 
 
 class OptionalFieldReduction(Reduction):
@@ -263,8 +346,46 @@ class OptionalFieldReduction(Reduction):
     def _finalize(bases, cuda=False, **kwargs):
         return xr.DataArray(bases[0], **kwargs)
 
-class count(OptionalFieldReduction):
-    """Count elements in each bin, returning the result as a uint32.
+
+class SelfIntersectingOptionalFieldReduction(OptionalFieldReduction):
+    """
+    Base class for optional field reductions for which self-intersecting
+    geometry may or may not be desireable.
+    Ignored if not using antialiasing.
+    """
+    def __init__(self, column=None, self_intersect=True):
+        super().__init__(column)
+        self.self_intersect = self_intersect
+
+    def _antialias_requires_2_stages(self):
+        return not self.self_intersect
+
+    def _build_append(self, dshape, schema, cuda, antialias, self_intersect):
+        if antialias and not self_intersect:
+            # append functions specific to antialiased lines without self_intersect
+            if cuda:
+                if self.column is None:
+                    return self._append_no_field_antialias_cuda_not_self_intersect
+                else:
+                    return self._append_antialias_cuda_not_self_intersect
+            else:
+                if self.column is None:
+                    return self._append_no_field_antialias_not_self_intersect
+                else:
+                    return self._append_antialias_not_self_intersect
+
+        # Fall back to base class implementation
+        return super()._build_append(dshape, schema, cuda, antialias, self_intersect)
+
+    def _hashable_inputs(self):
+        # Reductions with different self_intersect attributes much have different hashes otherwise
+        # toolz.memoize will treat them as the same to give incorrect results.
+        return super()._hashable_inputs() + (self.self_intersect,)
+
+
+class count(SelfIntersectingOptionalFieldReduction):
+    """Count elements in each bin, returning the result as a uint32, or a
+    float32 if using antialiasing.
 
     Parameters
     ----------
@@ -272,26 +393,67 @@ class count(OptionalFieldReduction):
         If provided, only counts elements in ``column`` that are not ``NaN``.
         Otherwise, counts every element.
     """
-    _dshape = dshape(ct.uint32)
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(ct.float32) if antialias else dshape(ct.uint32)
+
+    def _antialias_stage_2(self, self_intersect, array_module):
+        if self_intersect:
+            return ((AntialiasCombination.SUM_1AGG, array_module.nan),)
+        else:
+            return ((AntialiasCombination.SUM_2AGG, array_module.nan),)
 
     # CPU append functions
-    @staticmethod
-    @ngjit
-    def _append_no_field(x, y, agg):
-        agg[y, x] += 1
-
-
     @staticmethod
     @ngjit
     def _append(x, y, agg, field):
         if not isnull(field):
             agg[y, x] += 1
 
+    @staticmethod
+    @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        if not isnull(field):
+            if isnull(agg[y, x]):
+                agg[y, x] = aa_factor
+            else:
+                agg[y, x] += aa_factor
+
+    @staticmethod
+    @ngjit
+    def _append_antialias_not_self_intersect(x, y, agg, field, aa_factor):
+        if not isnull(field):
+            if isnull(agg[y, x]) or aa_factor > agg[y, x]:
+                agg[y, x] = aa_factor
+
+    @staticmethod
+    @ngjit
+    def _append_no_field(x, y, agg):
+        agg[y, x] += 1
+
+    @staticmethod
+    @ngjit
+    def _append_no_field_antialias(x, y, agg, aa_factor):
+        if isnull(agg[y, x]):
+            agg[y, x] = aa_factor
+        else:
+            agg[y, x] += aa_factor
+
+    @staticmethod
+    @ngjit
+    def _append_no_field_antialias_not_self_intersect(x, y, agg, aa_factor):
+        if isnull(agg[y, x]) or aa_factor > agg[y, x]:
+            agg[y, x] = aa_factor
+
     # GPU append functions
     @staticmethod
     @nb_cuda.jit(device=True)
-    def _append_no_field_cuda(x, y, agg):
-        nb_cuda.atomic.add(agg, (y, x), 1)
+    def _append_antialias_cuda(x, y, agg, field, aa_factor):
+        cuda_atomic_nanmax(agg, (y, x), field*aa_factor)
+
+    @staticmethod
+    @nb_cuda.jit(device=True)
+    def _append_no_field_antialias_cuda_not_self_intersect(x, y, agg, aa_factor):
+        cuda_atomic_nanmax(agg, (y, x), aa_factor)
 
     @staticmethod
     @nb_cuda.jit(device=True)
@@ -300,12 +462,31 @@ class count(OptionalFieldReduction):
             nb_cuda.atomic.add(agg, (y, x), 1)
 
     @staticmethod
-    def _create(shape, array_module):
-        return array_module.zeros(shape, dtype='u4')
+    @nb_cuda.jit(device=True)
+    def _append_no_field_antialias_cuda(x, y, agg, aa_factor):
+        cuda_atomic_nanmax(agg, (y, x), aa_factor)
+
+    @staticmethod
+    @nb_cuda.jit(device=True)
+    def _append_no_field_cuda(x, y, agg):
+        nb_cuda.atomic.add(agg, (y, x), 1)
+
+    def _build_combine(self, dshape, antialias):
+        if antialias:
+            return self._combine_antialias
+        else:
+            return self._combine
 
     @staticmethod
     def _combine(aggs):
         return aggs.sum(axis=0, dtype='u4')
+
+    @staticmethod
+    def _combine_antialias(aggs):
+        ret = aggs[0]
+        for i in range(1, len(aggs)):
+            nansum_in_place(ret, aggs[i])
+        return ret
 
 
 class by(Reduction):
@@ -353,19 +534,25 @@ class by(Reduction):
         self.preprocess.validate(in_dshape)
         self.reduction.validate(in_dshape)
 
-    def out_dshape(self, input_dshape):
+    def out_dshape(self, input_dshape, antialias):
         cats = self.categorizer.categories(input_dshape)
-        red_shape = self.reduction.out_dshape(input_dshape)
+        red_shape = self.reduction.out_dshape(input_dshape, antialias)
         return dshape(Record([(c, red_shape) for c in cats]))
 
     @property
     def inputs(self):
         return (self.preprocess, )
 
-    def _build_create(self, out_dshape):
-        n_cats = len(out_dshape.measure.fields)
+    def _antialias_requires_2_stages(self):
+        return self.reduction._antialias_requires_2_stages()
+
+    def _antialias_stage_2(self, self_intersect, array_module):
+        return self.reduction._antialias_stage_2(self_intersect, array_module)
+
+    def _build_create(self, required_dshape):
+        n_cats = len(required_dshape.measure.fields)
         return lambda shape, array_module: self.reduction._build_create(
-            out_dshape)(shape + (n_cats,), array_module)
+            required_dshape)(shape + (n_cats,), array_module)
 
     def _build_bases(self, cuda=False):
         bases = self.reduction._build_bases(cuda)
@@ -373,11 +560,11 @@ class by(Reduction):
             return bases
         return tuple(by(self.categorizer, base) for base in bases)
 
-    def _build_append(self, dshape, schema, cuda=False):
-        return self.reduction._build_append(dshape, schema, cuda)
+    def _build_append(self, dshape, schema, cuda, antialias, self_intersect):
+        return self.reduction._build_append(dshape, schema, cuda, antialias, self_intersect)
 
-    def _build_combine(self, dshape):
-        return self.reduction._combine
+    def _build_combine(self, dshape, antialias):
+        return self.reduction._build_combine(dshape, antialias)
 
     def _build_finalize(self, dshape):
         cats = list(self.categorizer.categories(dshape))
@@ -395,35 +582,65 @@ class any(OptionalFieldReduction):
     Parameters
     ----------
     column : str, optional
-        If provided, only elements in ``column`` that are ``NaN`` are skipped.
+        If provided, any elements in ``column`` that are ``NaN`` are skipped.
     """
-    _dshape = dshape(ct.bool_)
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(ct.float32) if antialias else dshape(ct.bool_)
 
-    @staticmethod
-    @ngjit
-    def _append_no_field(x, y, agg):
-        agg[y, x] = True
-    _append_no_field_cuda = _append_no_field
+    def _antialias_stage_2(self, self_intersect, array_module):
+        return ((AntialiasCombination.MAX, array_module.nan),)
 
+    # CPU append functions
     @staticmethod
     @ngjit
     def _append(x, y, agg, field):
         if not isnull(field):
             agg[y, x] = True
-    _append_cuda =_append
 
     @staticmethod
-    def _create(shape, array_module):
-        return array_module.zeros(shape, dtype='bool')
+    @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        if not isnull(field):
+            if isnull(agg[y, x]) or aa_factor > agg[y, x]:
+                agg[y, x] = aa_factor
+
+    @staticmethod
+    @ngjit
+    def _append_no_field(x, y, agg):
+        agg[y, x] = True
+
+    @staticmethod
+    @ngjit
+    def _append_no_field_antialias(x, y, agg, aa_factor):
+        if isnull(agg[y, x]) or aa_factor > agg[y, x]:
+            agg[y, x] = aa_factor
+
+    # GPU append functions
+    _append_cuda =_append
+    _append_no_field_cuda = _append_no_field
+
+    def _build_combine(self, dshape, antialias):
+        if antialias:
+            return self._combine_antialias
+        else:
+            return self._combine
 
     @staticmethod
     def _combine(aggs):
         return aggs.sum(axis=0, dtype='bool')
 
+    @staticmethod
+    def _combine_antialias(aggs):
+        ret = aggs[0]
+        for i in range(1, len(aggs)):
+            nanmax_in_place(ret, aggs[i])
+        return ret
+
 
 class _upsample(Reduction):
     """"Special internal class used for upsampling"""
-    _dshape = dshape(Option(ct.float64))
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(Option(ct.float64))
 
     @staticmethod
     def _finalize(bases, cuda=False, **kwargs):
@@ -433,11 +650,10 @@ class _upsample(Reduction):
     def inputs(self):
         return (extract(self.column),)
 
-    @staticmethod
-    def _create(shape, array_module):
+    def _build_create(self, required_dshape):
         # Use uninitialized memory, the upsample function must explicitly set unused
         # values to nan
-        return array_module.empty(shape, dtype='f8')
+        return self._create_float64_empty
 
     @staticmethod
     @ngjit
@@ -446,7 +662,7 @@ class _upsample(Reduction):
         pass
 
     @staticmethod
-    @ngjit
+    @nb_cuda.jit(device=True)
     def _append_cuda(x, y, agg, field):
         # not called, the upsample function must set agg directly
         pass
@@ -458,11 +674,8 @@ class _upsample(Reduction):
 
 class FloatingReduction(Reduction):
     """Base classes for reductions that always have floating-point dtype."""
-    _dshape = dshape(Option(ct.float64))
-
-    @staticmethod
-    def _create(shape, array_module):
-        return array_module.full(shape, np.nan, dtype='f8')
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(Option(ct.float64))
 
     @staticmethod
     def _finalize(bases, cuda=False, **kwargs):
@@ -476,21 +689,43 @@ class _sum_zero(FloatingReduction):
     ----------
     column : str
         Name of the column to aggregate over. Column data type must be numeric.
-        ``NaN`` values in the column are skipped.
     """
+    def _antialias_stage_2(self, self_intersect, array_module):
+        if self_intersect:
+            return ((AntialiasCombination.SUM_1AGG, 0),)
+        else:
+            return ((AntialiasCombination.SUM_2AGG, 0),)
 
-    @staticmethod
-    def _create(shape, array_module):
-        return array_module.zeros(shape, dtype='f8')
+    def _build_create(self, required_dshape):
+        return self._create_float64_zero
 
+    # CPU append functions.
     @staticmethod
     @ngjit
     def _append(x, y, agg, field):
         if not isnull(field):
+            # agg[y, x] cannot be null as initialised to zero.
             agg[y, x] += field
 
     @staticmethod
     @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if not isnull(value):
+            # agg[y, x] cannot be null as initialised to zero.
+            agg[y, x] += value
+
+    @staticmethod
+    @ngjit
+    def _append_antialias_not_self_intersect(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if not isnull(value) and value > agg[y, x]:
+            # agg[y, x] cannot be null as initialised to zero.
+            agg[y, x] = value
+
+    # GPU append functions
+    @staticmethod
+    @nb_cuda.jit(device=True)
     def _append_cuda(x, y, agg, field):
         if not isnull(field):
             nb_cuda.atomic.add(agg, (y, x), field)
@@ -499,7 +734,39 @@ class _sum_zero(FloatingReduction):
     def _combine(aggs):
         return aggs.sum(axis=0, dtype='f8')
 
-class sum(FloatingReduction):
+
+class SelfIntersectingFloatingReduction(FloatingReduction):
+    """
+    Base class fo floating reductions for which self-intersecting geometry
+    may or may not be desireable.
+    Ignored if not using antialiasing.
+    """
+    def __init__(self, column=None, self_intersect=True):
+        super().__init__(column)
+        self.self_intersect = self_intersect
+
+    def _antialias_requires_2_stages(self):
+        return not self.self_intersect
+
+    def _build_append(self, dshape, schema, cuda, antialias, self_intersect):
+        if antialias and not self_intersect:
+            if cuda:
+                raise NotImplementedError("SelfIntersectingOptionalFieldReduction")
+            else:
+                if self.column is None:
+                    return self._append_no_field_antialias_not_self_intersect
+                else:
+                    return self._append_antialias_not_self_intersect
+
+        return super()._build_append(dshape, schema, cuda, antialias, self_intersect)
+
+    def _hashable_inputs(self):
+        # Reductions with different self_intersect attributes much have different hashes otherwise
+        # toolz.memoize will treat them as the same to give incorrect results.
+        return super()._hashable_inputs() + (self.self_intersect,)
+
+
+class sum(SelfIntersectingFloatingReduction):
     """Sum of all elements in ``column``.
 
     Elements of resulting aggregate are nan if they are not updated.
@@ -510,26 +777,19 @@ class sum(FloatingReduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
+    def _antialias_stage_2(self, self_intersect, array_module):
+        if self_intersect:
+            return ((AntialiasCombination.SUM_1AGG, array_module.nan),)
+        else:
+            return ((AntialiasCombination.SUM_2AGG, array_module.nan),)
 
-    # Cuda implementation
     def _build_bases(self, cuda=False):
         if cuda:
             return (_sum_zero(self.column), any(self.column))
         else:
             return (self,)
 
-    @staticmethod
-    def _finalize(bases, cuda=False, **kwargs):
-        if cuda:
-            sums, anys = bases
-            x = np.where(anys, sums, np.nan)
-            return xr.DataArray(x, **kwargs)
-        else:
-            return xr.DataArray(bases[0], **kwargs)
-
-    # Single pass CPU implementation
-    # These methods will only be called if _build_bases returned (self,)
+    # CPU append functions
     @staticmethod
     @ngjit
     def _append(x, y, agg, field):
@@ -540,8 +800,35 @@ class sum(FloatingReduction):
                 agg[y, x] += field
 
     @staticmethod
+    @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if not isnull(value):
+            if isnull(agg[y, x]):
+                agg[y, x] = value
+            else:
+                agg[y, x] += value
+
+    @staticmethod
+    @ngjit
+    def _append_antialias_not_self_intersect(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if not isnull(value):
+            if isnull(agg[y, x]) or value > agg[y, x]:
+                agg[y, x] = value
+
+    @staticmethod
     def _combine(aggs):
         return nansum_missing(aggs, axis=0)
+
+    @staticmethod
+    def _finalize(bases, cuda=False, **kwargs):
+        if cuda:
+            sums, anys = bases
+            x = np.where(anys, sums, np.nan)
+            return xr.DataArray(x, **kwargs)
+        else:
+            return xr.DataArray(bases[0], **kwargs)
 
 
 class m2(FloatingReduction):
@@ -556,19 +843,17 @@ class m2(FloatingReduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
-
-    @staticmethod
-    def _create(shape, array_module):
-        return array_module.full(shape, 0.0, dtype='f8')
-
-    def _build_temps(self, cuda=False):
-        return (_sum_zero(self.column), count(self.column))
-
-    def _build_append(self, dshape, schema, cuda=False):
+    def _build_append(self, dshape, schema, cuda, antialias, self_intersect):
         if cuda:
             raise ValueError("""\
 The 'std' and 'var' reduction operations are not yet supported on the GPU""")
-        return super(m2, self)._build_append(dshape, schema, cuda)
+        return super(m2, self)._build_append(dshape, schema, cuda, antialias, self_intersect)
+
+    def _build_create(self, required_dshape):
+        return self._create_float64_zero
+
+    def _build_temps(self, cuda=False):
+        return (_sum_zero(self.column), count(self.column))
 
     @staticmethod
     @ngjit
@@ -597,6 +882,13 @@ class min(FloatingReduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
+    def _antialias_requires_2_stages(self):
+        return True
+
+    def _antialias_stage_2(self, self_intersect, array_module):
+        return ((AntialiasCombination.MIN, array_module.nan),)
+
+    # CPU append functions
     @staticmethod
     @ngjit
     def _append(x, y, agg, field):
@@ -607,6 +899,14 @@ class min(FloatingReduction):
 
     @staticmethod
     @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if isnull(agg[y, x]) or value > agg[y, x]:
+            agg[y, x] = value
+
+    # GPU append functions
+    @staticmethod
+    @nb_cuda.jit(device=True)
     def _append_cuda(x, y, agg, field):
         cuda_atomic_nanmin(agg, (y, x), field)
 
@@ -624,6 +924,10 @@ class max(FloatingReduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
+    def _antialias_stage_2(self, self_intersect, array_module):
+        return ((AntialiasCombination.MAX, array_module.nan),)
+
+    # CPU append functions
     @staticmethod
     @ngjit
     def _append(x, y, agg, field):
@@ -634,6 +938,19 @@ class max(FloatingReduction):
 
     @staticmethod
     @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if isnull(agg[y, x]) or value > agg[y, x]:
+            agg[y, x] = value
+
+    # GPU append functions
+    @staticmethod
+    @nb_cuda.jit(device=True)
+    def _append_antialias_cuda(x, y, agg, field, aa_factor):
+        cuda_atomic_nanmax(agg, (y, x), field*aa_factor)
+
+    @staticmethod
+    @nb_cuda.jit(device=True)
     def _append_cuda(x, y, agg, field):
         cuda_atomic_nanmax(agg, (y, x), field)
 
@@ -666,8 +983,6 @@ class mean(Reduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
-
     def _build_bases(self, cuda=False):
         return (_sum_zero(self.column), count(self.column))
 
@@ -688,8 +1003,6 @@ class var(Reduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
-
     def _build_bases(self, cuda=False):
         return (_sum_zero(self.column), count(self.column), m2(self.column))
 
@@ -710,8 +1023,6 @@ class std(Reduction):
         Name of the column to aggregate over. Column data type must be numeric.
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
-
     def _build_bases(self, cuda=False):
         return (_sum_zero(self.column), count(self.column), m2(self.column))
 
@@ -737,23 +1048,35 @@ class first(Reduction):
         Name of the column to aggregate over. If the data type is floating point,
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(Option(ct.float64))
+
+    def _antialias_requires_2_stages(self):
+        return True
+
+    def _antialias_stage_2(self, self_intersect, array_module):
+        return ((AntialiasCombination.FIRST, array_module.nan),)
 
     @staticmethod
-    def _append(x, y, agg):
-        raise NotImplementedError("first is currently implemented only for rasters")
+    @ngjit
+    def _append(x, y, agg, field):
+        if not isnull(field) and isnull(agg[y, x]):
+            agg[y, x] = field
 
     @staticmethod
-    def _create(shape, array_module):
-        raise NotImplementedError("first is currently implemented only for rasters")
+    @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if isnull(agg[y, x]) or value > agg[y, x]:
+            agg[y, x] = value
 
     @staticmethod
     def _combine(aggs):
-        raise NotImplementedError("first is currently implemented only for rasters")
+        raise NotImplementedError("first is not implemented for dask DataFrames")
 
     @staticmethod
-    def _finalize(bases, **kwargs):
-        raise NotImplementedError("first is currently implemented only for rasters")
+    def _finalize(bases, cuda=False, **kwargs):
+        return xr.DataArray(bases[0], **kwargs)
 
 
 
@@ -771,23 +1094,35 @@ class last(Reduction):
         Name of the column to aggregate over. If the data type is floating point,
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(Option(ct.float64))
+
+    def _antialias_requires_2_stages(self):
+        return True
+
+    def _antialias_stage_2(self, self_intersect, array_module):
+        return ((AntialiasCombination.LAST, array_module.nan),)
 
     @staticmethod
-    def _append(x, y, agg):
-        raise NotImplementedError("last is currently implemented only for rasters")
+    @ngjit
+    def _append(x, y, agg, field):
+        if not isnull(field):
+            agg[y, x] = field
 
     @staticmethod
-    def _create(shape, array_module):
-        raise NotImplementedError("last is currently implemented only for rasters")
+    @ngjit
+    def _append_antialias(x, y, agg, field, aa_factor):
+        value = field*aa_factor
+        if isnull(agg[y, x]) or value > agg[y, x]:
+            agg[y, x] = value
 
     @staticmethod
     def _combine(aggs):
-        raise NotImplementedError("last is currently implemented only for rasters")
+        raise NotImplementedError("last is not implemented for dask DataFrames")
 
     @staticmethod
-    def _finalize(bases, **kwargs):
-        raise NotImplementedError("last is currently implemented only for rasters")
+    def _finalize(bases, cuda=False, **kwargs):
+        return xr.DataArray(bases[0], **kwargs)
 
 
 
@@ -808,14 +1143,11 @@ class mode(Reduction):
         Name of the column to aggregate over. If the data type is floating point,
         ``NaN`` values in the column are skipped.
     """
-    _dshape = dshape(Option(ct.float64))
+    def out_dshape(self, in_dshape, antialias):
+        return dshape(Option(ct.float64))
 
     @staticmethod
     def _append(x, y, agg):
-        raise NotImplementedError("mode is currently implemented only for rasters")
-
-    @staticmethod
-    def _create(shape, array_module):
         raise NotImplementedError("mode is currently implemented only for rasters")
 
     @staticmethod
@@ -841,6 +1173,15 @@ class summary(Expr):
 
     >>> import datashader as ds
     >>> red = ds.summary(mean_a=ds.mean('a'), sum_b=ds.sum('b'))
+
+    Notes
+    -----
+    A single pass of the source dataset using antialiased lines can either be
+    performed using a single-stage aggregation (e.g. ``self_intersect=True``)
+    or two stages (``self_intersect=False``). If a ``summary`` contains a
+    ``count`` or ``sum`` reduction with ``self_intersect=False``, or any of
+    ``first``, ``last`` or ``min``, then the antialiased line pass will be
+    performed in two stages.
     """
     def __init__(self, **kwargs):
         ks, vs = zip(*sorted(kwargs.items()))
@@ -854,14 +1195,9 @@ class summary(Expr):
         for v in self.values:
             v.validate(input_dshape)
 
-    def out_dshape(self, in_dshape):
-        return dshape(Record([(k, v.out_dshape(in_dshape)) for (k, v)
-                              in zip(self.keys, self.values)]))
-
     @property
     def inputs(self):
         return tuple(unique(concat(v.inputs for v in self.values)))
-
 
 
 __all__ = list(set([_k for _k,_v in locals().items()
