@@ -1,4 +1,4 @@
-from __future__ import absolute_import
+from __future__ import annotations
 from datashader.colors import rgb, hex_to_rgb
 
 import pytest
