@@ -1,4 +1,4 @@
-from __future__ import division, absolute_import
+from __future__ import annotations
 
 from io import BytesIO
 
@@ -25,7 +25,10 @@ def build_agg(array_module=np):
     c = array_module.arange(10, 19, dtype='f8').reshape((3, 3))
     c[[0, 1, 2], [0, 1, 2]] = array_module.nan
     s_c = xr.DataArray(c, coords=coords, dims=dims)
-    agg = xr.Dataset(dict(a=s_a, b=s_b, c=s_c))
+    d = array_module.arange(10, 19, dtype='u4').reshape((3, 3))
+    d[[0, 1, 2, 2], [0, 1, 2, 1]] = 1
+    s_d = xr.DataArray(d, coords=coords, dims=dims)
+    agg = xr.Dataset(dict(a=s_a, b=s_b, c=s_c, d=s_d))
     return agg
 
 
@@ -71,13 +74,19 @@ solutions = {how: tf.Image(np.array(v, dtype='u4'),
                            coords=coords, dims=dims)
              for how, v in solution_lists.items()}
 
-eq_hist_sol = {'a': np.array([[0, 4291543295, 4288846335],
-                              [4286149631, 0, 4283518207],
-                              [4280821503, 4278190335, 0]], dtype='u4'),
-               'b': np.array([[0, 4291543295, 4288846335],
-                              [4286609919, 0, 4283518207],
-                              [4281281791, 4278190335, 0]], dtype='u4')}
-eq_hist_sol['c'] = eq_hist_sol['b']
+# Same result obtained regardless of data dtype (u4, f4, f8)
+eq_hist_sol = np.array([[0, 4291543295, 4288846335],
+                        [4286149631, 0, 4283518207],
+                        [4280821503, 4278190335, 0]], dtype='u4')
+
+eq_hist_sol_rescale_discrete_levels = {
+    'a': np.array([[0, 4289306879, 4287070463],
+                   [4284834047, 0, 4282597631],
+                   [4280361215, 4278190335, 0]], dtype='u4'),
+    'b': np.array([[0, 4289306879, 4287070207],
+                   [4284834047, 0, 4282597375],
+                   [4280361215, 4278190335, 0]], dtype='u4')}
+eq_hist_sol_rescale_discrete_levels['c'] = eq_hist_sol_rescale_discrete_levels['b']
 
 
 def check_span(x, cmap, how, sol):
@@ -153,9 +162,19 @@ def test_shade(agg, attr, span):
     assert_eq_xr(img, sol)
 
     # span option not supported with how='eq_hist'
-    img = tf.shade(x, cmap=cmap, how='eq_hist')
-    sol = tf.Image(eq_hist_sol[attr], coords=coords, dims=dims)
-    assert_eq_xr(img, sol)
+    if span is None:
+        img = tf.shade(x, cmap=cmap, how='eq_hist', rescale_discrete_levels=False)
+        sol = tf.Image(eq_hist_sol, coords=coords, dims=dims)
+        assert_eq_xr(img, sol)
+
+        img = tf.shade(x, cmap=cmap, how='eq_hist', rescale_discrete_levels=True)
+        sol = tf.Image(eq_hist_sol_rescale_discrete_levels[attr], coords=coords, dims=dims)
+        if cupy and attr=='a' and isinstance(agg.a.data, cupy.ndarray):
+            # cupy eq_hist has slightly different numerics hence slightly different RGBA results
+            sol = sol.copy(deep=True)
+            sol[2, 0] = sol[2, 0] - 0x100
+
+        assert_eq_xr(img, sol)
 
     img = tf.shade(x, cmap=cmap,
                    how=lambda x, mask: np.where(mask, np.nan, x ** 2))
@@ -461,6 +480,7 @@ def test_shade_category(array):
     assert ((img.data[1,0] >> 24) & 0xFF) == 20 # min alpha
     assert ((img.data[1,1] >> 24) & 0xFF) == 20 # min alpha
 
+
 @pytest.mark.parametrize('array', arrays)
 def test_shade_zeros(array):
     coords = [np.array([0, 1]), np.array([2, 5])]
@@ -474,6 +494,26 @@ def test_shade_zeros(array):
     img = tf.shade(cat_agg, color_key=colors, how='linear', min_alpha=0)
     sol = np.array([[5584810, 5584810],
                     [5584810, 5584810]], dtype='u4')
+    sol = tf.Image(sol, coords=coords, dims=dims)
+    assert_eq_xr(img, sol)
+
+
+@pytest.mark.parametrize('agg', aggs)
+@pytest.mark.parametrize('attr', ['d'])
+@pytest.mark.parametrize('rescale', [False, True])
+#@pytest.mark.parametrize('rescale', [True])
+def test_shade_rescale_discrete_levels(agg, attr, rescale):
+    x = getattr(agg, attr)
+    cmap = ['pink', 'red']
+    img = tf.shade(x, cmap=cmap, how='eq_hist', rescale_discrete_levels=rescale)
+    if rescale:
+        sol = np.array([[0xff8981ff, 0xff6d67ff, 0xff524dff],
+                        [0xff3633ff, 0xff8981ff, 0xff1b19ff],
+                        [0xff0000ff, 0xff8981ff, 0xff8981ff]], dtype='uint32')
+    else:
+        sol = np.array([[0xffcbc0ff, 0xffa299ff, 0xff7973ff],
+                        [0xff514cff, 0xffcbc0ff, 0xff2826ff],
+                        [0xff0000ff, 0xffcbc0ff, 0xffcbc0ff]], dtype='uint32')
     sol = tf.Image(sol, coords=coords, dims=dims)
     assert_eq_xr(img, sol)
 
@@ -929,7 +969,7 @@ def test_int_array_density():
     assert np.allclose(tf._array_density(data, float_type=False), 0.75)
     assert np.allclose(tf._array_density(data, float_type=False, px=3), 1)
 
-    
+
 def test_float_array_density():
     data = np.ones((4, 4), dtype='float32')
     assert tf._array_density(data, float_type=True) == 1.0
@@ -940,7 +980,7 @@ def test_float_array_density():
     data[2, 0] = data[0, 2] = data[1, 1] = 1
     assert np.allclose(tf._array_density(data, float_type=True), 0.75)
     assert np.allclose(tf._array_density(data, float_type=True, px=3), 1)
-    
+
 
 def test_rgb_dynspread():
     b = 0xffff0000
@@ -1034,13 +1074,13 @@ def test_eq_hist():
     data[np.random.randint(300**2, size=100)] = np.nan
     data = (data - np.nanmin(data)).reshape((300, 300))
     mask = np.isnan(data)
-    eq = tf.eq_hist(data, mask)
+    eq, _ = tf.eq_hist(data, mask)
     check_eq_hist_cdf_slope(eq)
     assert (np.isnan(eq) == mask).all()
     # Integer
     data = np.random.normal(scale=100, size=(300, 300)).astype('i8')
     data = data - data.min()
-    eq = tf.eq_hist(data)
+    eq, _ = tf.eq_hist(data)
     check_eq_hist_cdf_slope(eq)
 
 
@@ -1064,3 +1104,43 @@ def test_shade_should_handle_zeros_array():
     arr = tf.Image(data, dims=['x', 'y'])
     img = tf.shade(arr, cmap=['white', 'black'], how='linear')
     assert img is not None
+
+
+def test_shade_with_discrete_color_key():
+    data = np.array([[0, 0, 0, 0, 0],
+                     [0, 1, 1, 1, 0],
+                     [0, 2, 2, 2, 0],
+                     [0, 3, 3, 3, 0],
+                     [0, 0, 0, 0, 0]], dtype='uint32')
+    color_key = {1: 'white', 2: 'purple', 3: 'yellow'}
+    result = np.array([[0, 0, 0, 0, 0],
+                       [0, 4294967295, 4294967295, 4294967295, 0],
+                       [0, 4286578816, 4286578816, 4286578816, 0],
+                       [0, 4278255615, 4278255615, 4278255615, 0],
+                       [0, 0, 0, 0, 0]],
+                      dtype='uint32')
+
+    # numpy case
+    arr_numpy = tf.Image(data, dims=['x', 'y'])
+    result_numpy = tf.shade(arr_numpy, color_key=color_key)
+    assert (result_numpy.data == result).all()
+
+    # dask with numpy backed case
+    arr_dask = tf.Image(da.from_array(data, chunks=(2, 2)), dims=['x', 'y'])
+    result_dask = tf.shade(arr_dask, color_key=color_key)
+    assert (result_dask.data == result).all()
+
+    # cupy case
+    try:
+        import cupy
+        arr_cupy = tf.Image(cupy.asarray(data), dims=['x', 'y'])
+        result_cupy = tf.shade(arr_cupy, color_key=color_key)
+        assert (result_cupy.data == result).all()
+    except ImportError:
+        cupy = None
+
+
+def test_interpolate_alpha_discrete_levels_None():
+    data = np.array([[0.0, 1.0], [1.0, 0.0]])
+    # Issue #1084: this raises a ValueError.
+    tf._interpolate_alpha(data, data, None, "eq_hist", 0.5, None, 0.4, True)
